feat: restructure repository into monorepo
- Move all existing code to engine/ subdirectory
- Rename package from 'pyrat' to 'pyrat-engine'
- Update all imports and references
- Add top-level Makefile for monorepo orchestration
- Update CI/CD workflows for new structure
- Update pre-commit hooks for monorepo structure
- Add placeholder directories for future components (gui, protocol, examples, cli)
- Update documentation to reflect monorepo structure

BREAKING CHANGE: Package renamed from 'pyrat' to 'pyrat-engine'
