+++ .github/workflows/ci.yml
@@ -25,31 +25,31 @@ jobs:
       uses: actions/cache@v4
       with:
         path: ~/.cargo/registry
-        key: ${{ runner.os }}-cargo-registry-${{ hashFiles('**/Cargo.lock') }}
+        key: ${{ runner.os }}-cargo-registry-${{ hashFiles('engine/Cargo.lock') }}
 
     - name: Cache cargo index
       uses: actions/cache@v4
       with:
         path: ~/.cargo/git
-        key: ${{ runner.os }}-cargo-index-${{ hashFiles('**/Cargo.lock') }}
+        key: ${{ runner.os }}-cargo-index-${{ hashFiles('engine/Cargo.lock') }}
 
     - name: Cache cargo build
       uses: actions/cache@v4
       with:
-        path: target
-        key: ${{ runner.os }}-cargo-build-target-${{ hashFiles('**/Cargo.lock') }}
+        path: engine/target
+        key: ${{ runner.os }}-cargo-build-target-${{ hashFiles('engine/Cargo.lock') }}
 
     - name: Check formatting
-      run: cargo fmt --all -- --check
+      run: cd engine && cargo fmt --all -- --check
 
     - name: Run clippy
-      run: cargo clippy --all-targets --all-features -- -D warnings -A non-local-definitions
+      run: cd engine && cargo clippy --all-targets --all-features -- -D warnings -A non-local-definitions
 
     - name: Run tests
-      run: cargo test --verbose
+      run: cd engine && cargo test --verbose
 
     - name: Build release
-      run: cargo build --release --verbose
+      run: cd engine && cargo build --release --verbose
 
   python-tests:
     name: Python Tests
@@ -73,31 +73,32 @@ jobs:
       uses: actions/cache@v4
       with:
         path: ~/.cargo/registry
-        key: ${{ runner.os }}-cargo-registry-${{ hashFiles('**/Cargo.lock') }}
+        key: ${{ runner.os }}-cargo-registry-${{ hashFiles('engine/Cargo.lock') }}
 
     - name: Cache cargo index
       uses: actions/cache@v4
       with:
         path: ~/.cargo/git
-        key: ${{ runner.os }}-cargo-index-${{ hashFiles('**/Cargo.lock') }}
+        key: ${{ runner.os }}-cargo-index-${{ hashFiles('engine/Cargo.lock') }}
 
     - name: Cache cargo build
       uses: actions/cache@v4
       with:
-        path: target
-        key: ${{ runner.os }}-cargo-build-target-${{ hashFiles('**/Cargo.lock') }}
+        path: engine/target
+        key: ${{ runner.os }}-cargo-build-target-${{ hashFiles('engine/Cargo.lock') }}
 
     - name: Install uv
       uses: astral-sh/setup-uv@v6
       with:
         enable-cache: true
-        cache-dependency-glob: "pyproject.toml"
+        cache-dependency-glob: "engine/pyproject.toml"
 
     - name: Create virtual environment
-      run: uv venv .venv
+      run: cd engine && uv venv .venv
 
     - name: Build and install package with dependencies
       run: |
+        cd engine
         source .venv/bin/activate
         # Install the package with all dependencies (base + dev)
         uv pip install -e ".[dev]"
@@ -106,5 +107,6 @@ jobs:
 
     - name: Run Python tests
       run: |
+        cd engine
         source .venv/bin/activate
         pytest python/tests -v
