+++ .pre-commit-config.yaml
@@ -17,17 +17,17 @@ repos:
         -   id: ruff
             args: [--fix]
             types_or: [python, pyi]
-            files: ^python/
+            files: ^engine/python/
         -   id: ruff-format
             types_or: [python, pyi]
-            files: ^python/
+            files: ^engine/python/
 
   # Python type checking
   -   repo: https://github.com/pre-commit/mirrors-mypy
       rev: v1.8.0
       hooks:
         -   id: mypy
-            files: ^python/
+            files: ^engine/python/
             additional_dependencies:
               - gymnasium>=0.29.0
               - pettingzoo>=1.24.0
@@ -38,7 +38,7 @@ repos:
         -   id: fmt
             name: cargo fmt
             description: Format Rust code
-            entry: cargo fmt --all
+            entry: bash -c 'cd engine && cargo fmt --all'
             language: system
             types: [rust]
             pass_filenames: false
@@ -46,7 +46,7 @@ repos:
         -   id: clippy
             name: cargo clippy
             description: Lint Rust code
-            entry: cargo clippy --all-targets --all-features -- -D warnings -A non-local-definitions
+            entry: bash -c 'cd engine && cargo clippy --all-targets --all-features -- -D warnings -A non-local-definitions'
             language: system
             types: [rust]
             pass_filenames: false
@@ -56,7 +56,7 @@ repos:
       hooks:
         -   id: pytest
             name: pytest
-            entry: pytest
+            entry: bash -c 'cd engine && source .venv/bin/activate && pytest python/tests'
             language: system
             pass_filenames: false
             always_run: true
