+++ engine/python/pyrat_engine/_rust.pyi
