+++ engine/rust/src/bindings/game.rs
