+++ engine/rust/src/game/maze_generation.rs
